+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed formatting of `resourcestring` sections.
+- Fixed indentation inside `initialization`/`finalization` sections.
 
 ## [0.5.0] - 2025-03-31